# apd: remove some allocs from Decimal.setString