@nvanbenschoten 2022-06-14
CI is failing on old Go releases due to https://github.com/dominikh/go-tools/issues/1251. I'll fix that separately in https://github.com/cockroachdb/apd/pull/117.