@josharian 2022-06-14
> @josharian out of curiosity, are you using this library?

Yes, albeit indirectly, via [bojanz/currency](http://github.com/bojanz/currency).

> @josharian could you rebase this on master? Then I'll run it through CI again and merge on green.

Done. Thanks!